Spring Boot Rest Api Request Param

+++ src/main/java/springbootapi/springbootapiproject/Bean/Student.java
@@ -3,21 +3,21 @@ package springbootapi.springbootapiproject.Bean;
 
 
 public class Student {
-    private Long id;
+    private int id;
     private String firstName;
     private String lastName;
 
-    public Student(Long id, String firstName, String lastName) {
+    public Student(int id, String firstName, String lastName) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
